Adds init func and table stats faux func for DHCP information
This PR introduces dt_dhcp_init() and dt_dhcp_tbl_stats() the former
of which currently only adds a packet process entry for DHCP offer
packets and logs information to a file. The latter is a place holder
for how future implementation should be done where DHCP information
is collected, added to a table, and then shipped via the sender. Both
functions include a comment indicating their incomplete nature and a
date marking this commit.