Add DHCP offer packet processing entry and log
This commit introduces a new file dhcp_collector.c and the basics for the
collection of DHCP information in the device telemetry section of unum-sdk.
With this commit, current functionality only supports processing a DHCP
offer packet and logging from the processing entry callback which is not
considered good form. This commit is merely to establish the base for the
future work of moving the DHCP information collection out of the finger
printing section of the SDK and into the device telemetry section.